bolt lifespans can now be decimals

+++ src/main/java/com/gmail/thelilchicken01/ethermist/enchantment/custom_enchants/EnduringMagicEnchant.java
@@ -18,7 +18,7 @@ public record EnduringMagicEnchant() implements EnchantmentValueEffect {
         return CODEC;
     }
 
-    public static int modifyLifespan(int level, int lifespan) {
+    public static double modifyLifespan(int level, double lifespan) {
         return lifespan + (level);
     }
 }
